Rename `build` to `build_pdfium`
To avoid conflicts with the Python build module
